pyproject.toml
[project]
name = "aclarknet"
version = "0.1.0"
dependencies = [
  "babel",
  "boto3",
  "crispy-bootstrap5",
  "django-allauth",
  "django-crispy-forms",
  "django-extensions",
  "django-import-export",
  "django-mongodb-backend",
  "django-phonenumber-field",
  "django-recaptcha",
  "django-ses",
  "gunicorn",
  "html2docx",
  "phonenumbers",
  "python-docx",
  "python-dotenv",
  "python-webpack-boilerplate",
  "texttable",
  "wagtail",
  "xhtml2pdf",
]

[project.scripts]
dm = "dm:main"

[project.optional-dependencies]
dev = [
  "Faker",
  "django-debug-toolbar",
  "django-hijack",
  "prek",
  "sphinx",
  "furo",
  "myst-parser",
]
docs = [
  "sphinx",
  "furo",
  "myst-parser",
]
test = [
  "pytest",
  "pytest-django",
]

[tool.setuptools]
packages = ["aclarknet", "db"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "aclarknet.settings.dev"
